feat: Add token and message fields to APIKeys class
Add the 'token' and 'message' fields to the APIKeys class in the 'end_points.dart' file. This allows for easier access to these values when making API requests.

Also, update the 'AppColors' class in 'app_colors.dart' to include a new color, 'green'.

Update the English and Arabic translation files to include a new translation for the 'loginSuccess' key.

In the 'LoginModel' class in 'login_model.dart', update the 'message' field to use the 'APIKeys.message' key instead of 'APIKeys.email'.

Add a new file, 'custom_loading_indicator.dart', which contains a custom loading indicator widget.

In the 'DioConsumer' class in 'dio_consumer.dart', update the 'put' method to use the 'data' parameter instead of 'body'.

In the 'AuthRepository' class in 'auth_repository.dart', update the 'post' method to use the 'data' parameter instead of 'body'.

In the 'LoginCubit' class in 'login_cubit.dart', add new state classes for loading, error, and success states.

In the 'SignUpScreen' class in 'sign_up_screen.dart', update the 'listener' method to handle the new state classes.

In the 'commons.dart' file, add new methods for showing toast messages and define an enum for toast states.

In the 'ApiInterceptors' class in 'api_interceptors.dart', implement the 'onResponse' and 'onError' methods.

In the 'ApiConsumer' class in 'api_consumer.dart', update the 'post', 'delete', 'patch', and 'put' methods to include a 'data' parameter.

In the 'service_locator.dart' file, register the 'AuthRepository', 'ApiConsumer', and 'Dio' classes with the service locator.